ualberta-carpool-backend/ualberta_carpool_backend/api/serializers.py
from rest_framework import serializers
from .models import Driver, Passenger, Ride

class DriverSerializer(serializers.ModelSerializer):
    class Meta:
        model = Driver
        fields = "__all__"

class PassengerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Passenger
        fields = "__all__"

class RideSerializer(serializers.ModelSerializer):
    class Meta:
        model = Ride
        fields = "__all__"