ualberta-carpool-backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
phonenumbers==8.12.55
pytz==2022.2.1
sqlparse==0.4.2
